Organize all the llm stuff into a subpackage (#3436)
