+++ autogpt/llm/chat.py
@@ -3,10 +3,10 @@ from random import shuffle
 
 from openai.error import RateLimitError
 
-from autogpt import token_counter
-from autogpt.api_manager import ApiManager
 from autogpt.config import Config
-from autogpt.llm_utils import create_chat_completion
+from autogpt.llm.api_manager import ApiManager
+from autogpt.llm.llm_utils import create_chat_completion
+from autogpt.llm.token_counter import count_message_tokens
 from autogpt.logs import logger
 from autogpt.types.openai import Message
 
@@ -43,7 +43,7 @@ def generate_context(prompt, relevant_memory, full_message_history, model):
     next_message_to_add_index = len(full_message_history) - 1
     insertion_index = len(current_context)
     # Count the currently used tokens
-    current_tokens_used = token_counter.count_message_tokens(current_context, model)
+    current_tokens_used = count_message_tokens(current_context, model)
     return (
         next_message_to_add_index,
         current_tokens_used,
@@ -114,7 +114,7 @@ def chat_with_ai(
                     prompt, relevant_memory, full_message_history, model
                 )
 
-            current_tokens_used += token_counter.count_message_tokens(
+            current_tokens_used += count_message_tokens(
                 [create_chat_message("user", user_input)], model
             )  # Account for user input (appended later)
 
@@ -122,9 +122,7 @@ def chat_with_ai(
                 # print (f"CURRENT TOKENS USED: {current_tokens_used}")
                 message_to_add = full_message_history[next_message_to_add_index]
 
-                tokens_to_add = token_counter.count_message_tokens(
-                    [message_to_add], model
-                )
+                tokens_to_add = count_message_tokens([message_to_add], model)
                 if current_tokens_used + tokens_to_add > send_token_limit:
                     break
 
@@ -175,7 +173,7 @@ def chat_with_ai(
                 )
                 if not plugin_response or plugin_response == "":
                     continue
-                tokens_to_add = token_counter.count_message_tokens(
+                tokens_to_add = count_message_tokens(
                     [create_chat_message("system", plugin_response)], model
                 )
                 if current_tokens_used + tokens_to_add > send_token_limit:
