+++ autogpt/llm/llm_utils.py
@@ -8,8 +8,8 @@ import openai
 from colorama import Fore, Style
 from openai.error import APIError, RateLimitError, Timeout
 
-from autogpt.api_manager import ApiManager
 from autogpt.config import Config
+from autogpt.llm.api_manager import ApiManager
 from autogpt.logs import logger
 from autogpt.types.openai import Message
