+++ tests/conftest.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import pytest
 from pytest_mock import MockerFixture
 
-from autogpt.api_manager import ApiManager
 from autogpt.config import Config
+from autogpt.llm import ApiManager
 from autogpt.workspace import Workspace
 
 pytest_plugins = ["tests.integration.agent_factory"]
