+++ tests/test_agent.py
@@ -3,10 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from autogpt.agent import Agent
-from autogpt.chat import chat_with_ai
 from autogpt.config import Config
-from autogpt.speech import say_text
-from autogpt.utils import clean_input
 
 
 @pytest.fixture
